+++ src/components/DownloadComponents/helpers/downloadLessonResources.test.tsx
@@ -1,3 +1,5 @@
+import type { ResourcesToDownloadArrayType } from "../downloads.types";
+
 import downloadLessonResources from "./downloadLessonResources";
 import createDownloadResourcesLink from "./createDownloadResourcesLink";
 
@@ -16,11 +18,10 @@ const successResponse = {
   ok: true,
 };
 
-const resourcesToDownload = {
-  "exit-quiz-answers": true,
-  "worksheet-pdf": true,
-};
-
+const resourcesToDownload: ResourcesToDownloadArrayType = [
+  "exit-quiz-answers",
+  "worksheet-pdf",
+];
 const resourcesToDownloadAsSelection = "exit-quiz-answers,worksheet-pdf";
 
 describe("downloadLessonResources", () => {
@@ -30,7 +31,7 @@ describe("downloadLessonResources", () => {
 
   it("should return when no resource types are passed", async () => {
     console.log = jest.fn();
-    await downloadLessonResources("lesson-slug", {});
+    await downloadLessonResources("lesson-slug", []);
 
     expect(console.log).toHaveBeenCalledWith("no resources to download");
   });