+++ src/components/Sanity/TextBlock/TextBlockCardImageCta.tsx
@@ -44,7 +44,7 @@ const TextBlockCardImageCta: FC<
         $ph={[16, 0]}
         {...imageContainerProps}
       >
-        <CardImage imageSrc={image.imageSrc} alt={image.alt} />
+        <CardImage {...image} />
       </Flex>
       <Flex $flexDirection={"column"} $font={["body-2", "body-1"]}>
         <Heading $mb={32} $font={["heading-5", "heading-4"]} tag={"h2"}>